+++ src/nom_parser.rs
@@ -364,7 +364,10 @@ mod acceptance_tests {
 
     #[rstest]
     #[case("2023-02-27 txn\n  Assets:A 10 CHF", None)]
-    #[ignore = "not implemented"]
+    #[case(
+        "2023-02-27 txn\n  Assets:A 10 CHF @ 19 EUR",
+        Some((PriceType::Unit, Amount::new(19, "EUR")))
+    )]
     #[case(
         "2023-02-27 txn\n  Assets:A 10 CHF@19 EUR",
         Some((PriceType::Unit, Amount::new(19, "EUR")))