+++ src/transaction/posting.rs
@@ -227,7 +227,7 @@ pub fn posting(input: Span<'_>) -> IResult<'_, Posting<'_>> {
             preceded(space0, char('}')),
         ),
     ))(input)?;
-    let (input, price) = opt(preceded(space1, price))(input)?;
+    let (input, price) = opt(preceded(space0, price))(input)?;
     let (input, _) = space0(input)?;
     let (input, comment) = opt(comment)(input)?;
     Ok((
@@ -249,7 +249,35 @@ fn price(input: Span<'_>) -> IResult<'_, (PriceType, Amount<'_>)> {
             map(tag("@@"), |_| PriceType::Total),
             map(char('@'), |_| PriceType::Unit),
         )),
-        space1,
+        space0,
         amount,
     )(input)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[rstest]
+    #[case("Assets:A 10 CHF", None)]
+    #[case(
+        "Assets:A 10 CHF @ 19 EUR",
+        Some((PriceType::Unit, Amount::new(19, "EUR")))
+    )]
+    #[case(
+        "Assets:A 10 CHF@19 EUR",
+        Some((PriceType::Unit, Amount::new(19, "EUR")))
+    )]
+    #[case(
+        "Assets:A 10 CHF  @  2 EUR",
+        Some((PriceType::Unit, Amount::new(2, "EUR")))
+    )]
+    #[case(
+        "Assets:A 10 CHF  @@  20 EUR",
+        Some((PriceType::Total, Amount::new(20, "EUR")))
+    )]
+    fn parse_posting_price(#[case] input: &str, #[case] expected: Option<(PriceType, Amount<'_>)>) {
+        let (_, posting) = posting(Span::new(input)).unwrap();
+        assert_eq!(posting.price().map(|(t, a)| (t, a.clone())), expected);
+    }
 }